add http shared tests, batch passing shared
