+++ packages/apollo-link-batch-http/src/__tests__/batchHttpLink.ts
@@ -1,9 +1,18 @@
 import { ApolloLink, execute, Observable, makePromise } from 'apollo-link';
-import { ApolloFetch } from 'apollo-fetch';
 import { print } from 'graphql';
+import * as fetchMock from 'fetch-mock';
 import gql from 'graphql-tag';
 
+import { sharedHttpTest } from './sharedHttpTests';
 import { BatchHttpLink } from '../batchHttpLink';
+
+const sampleQuery = gql`
+  query SampleQuery {
+    stub {
+      id
+    }
+  }
+`;
 
 const operation = {
   query: gql`
@@ -16,14 +25,49 @@ const operation = {
 };
 
 describe('BatchHttpLink', () => {
+  sharedHttpTest(
+    'BatchHttpLink',
+    httpArgs => {
+      const args = {
+        ...httpArgs,
+        batchInterval: 0,
+        batchMax: 1,
+      };
+      return new BatchHttpLink(args);
+    },
+    true,
+  );
+
   beforeAll(() => {
     jest.resetModules();
   });
+
+  const data = { data: { hello: 'world' } };
+  const data2 = { data: { hello: 'everyone' } };
+
+  let subscriber;
+
+  beforeEach(() => {
+    const makePromise = res =>
+      new Promise((resolve, reject) => setTimeout(() => resolve(res)));
+    fetchMock.post('begin:batch', makePromise([data, data2]));
+
+    const next = jest.fn();
+    const error = jest.fn();
+    const complete = jest.fn();
+
+    subscriber = {
+      next,
+      error,
+      complete,
+    };
+  });
+
   it('does not need any constructor arguments', () => {
     expect(() => new BatchHttpLink()).not.toThrow();
   });
 
-  it('should pass batchInterval and batchMax to BatchLink', () => {
+  it('should pass batchInterval, batchMax, and batchKey to BatchLink', () => {
     jest.mock('apollo-link-batch', () => ({
       BatchLink: jest.fn(),
     }));
@@ -31,124 +75,103 @@ describe('BatchHttpLink', () => {
     const BatchLink = require('apollo-link-batch').BatchLink;
     const LocalScopedLink = require('../batchHttpLink').BatchHttpLink;
 
+    const batchKey = () => 'hi';
+    const batchHandler = operations => Observable.of();
+
     const batch = new LocalScopedLink({
       batchInterval: 20,
       batchMax: 20,
+      batchKey,
+      batchHandler,
     });
 
-    const { batchInterval, batchMax } = BatchLink.mock.calls[0][0];
+    const {
+      batchInterval,
+      batchMax,
+      batchKey: batchKeyArg,
+    } = BatchLink.mock.calls[0][0];
     expect(batchInterval).toBe(20);
     expect(batchMax).toBe(20);
+    expect(batchKeyArg()).toEqual(batchKey());
   });
 
-  it('should pass printed operation to apollo fetch', done => {
-    const apolloFetch: any = operations => {
-      expect(Array.isArray(operations)).toBeTruthy();
-      expect(operations.length).toEqual(1);
-      expect(operations[0].query).toEqual(print(operation.query));
-      done();
-      return makePromise(Observable.of());
-    };
-    (apolloFetch as any).use = () => void 0;
-    (apolloFetch as any).useAfter = () => void 0;
-    (apolloFetch as any).batchUse = () => void 0;
-    (apolloFetch as any).batchUseAfter = () => void 0;
-
-    const link = ApolloLink.from([
-      new BatchHttpLink({
-        fetch: apolloFetch as ApolloFetch,
-      }),
-    ]);
-
-    execute(link, operation).subscribe({});
-  });
-
-  it("should call observer's error when apollo fetch returns an error", done => {
-    const error = new Error('Evans Hauser');
-
-    const apolloFetch: any = operations => {
-      expect(Array.isArray(operations)).toBeTruthy();
-      expect(operations.length).toEqual(1);
-      expect(operations[0].query).toEqual(print(operation.query));
-      return makePromise(
-        new Observable(observer => {
-          observer.error(error);
-        }),
-      );
-    };
-    (apolloFetch as any).use = () => void 0;
-    (apolloFetch as any).useAfter = () => void 0;
-    (apolloFetch as any).batchUse = () => void 0;
-    (apolloFetch as any).batchUseAfter = () => void 0;
-
-    const link = ApolloLink.from([
-      new BatchHttpLink({
-        fetch: apolloFetch as ApolloFetch,
-      }),
-    ]);
-
-    execute(link, operation).subscribe({
-      error: received => {
-        expect(received).toEqual(error);
-        done();
-      },
+  it('handles batched requests', done => {
+    const link = new BatchHttpLink({
+      uri: 'batch',
+      batchInterval: 0,
+      batchMax: 2,
     });
-  });
 
-  it("should call observer's next and then complete when apollo fetch returns data", done => {
-    const results = [
-      {
-        data: {
-          data: {
-            works: 'great',
-          },
-        },
-      },
-    ];
-
-    let middleware = [];
-    const apolloFetch: any = operations => {
-      expect(Array.isArray(operations)).toBeTruthy();
-      expect(operations.length).toEqual(1);
-      expect(operations[0].query).toEqual(print(operation.query));
-      middleware.forEach(x => x());
-      return makePromise(
-        new Observable(observer => {
-          observer.next(results);
-          observer.complete();
-        }),
-      );
+    let nextCalls = 0;
+    let completions = 0;
+    const next = expectedData => data => {
+      try {
+        expect(data).toEqual(expectedData);
+        nextCalls++;
+      } catch (error) {
+        done.fail(error);
+      }
     };
-    (apolloFetch as any).use = () => void 0;
-    (apolloFetch as any).useAfter = () => void 0;
-    (apolloFetch as any).batchUse = fn => {
-      const request = { options: { headers: {} } } as any;
-      const next = () => {
-        expect(request.options.headers.foo).toEqual(true);
-      };
-      middleware.push(() => fn(request, next));
+
+    const complete = () => {
+      try {
+        const calls = fetchMock.calls('begin:batch');
+        expect(calls.length).toBe(1);
+        expect(nextCalls).toBe(2);
+
+        const options = fetchMock.lastOptions('begin:batch');
+        expect(options.credentials).toEqual('two'); //test reduceFetchOptions
+
+        completions++;
+
+        if (completions === 2) {
+          done();
+        }
+      } catch (error) {
+        done.fail(error);
+      }
     };
-    (apolloFetch as any).batchUseAfter = () => void 0;
 
-    const link = ApolloLink.from([
-      new BatchHttpLink({
-        fetch: apolloFetch as ApolloFetch,
-      }),
-    ]);
-
-    const next = jest.fn();
+    const error = error => {
+      done.fail(error);
+    };
 
     execute(link, {
-      ...operation,
-      context: {
-        headers: { foo: true },
-      },
-    }).subscribe({
-      next,
-      complete: () => {
-        expect(next).toBeCalledWith(results[0]);
-        done();
-      },
+      query: sampleQuery,
+      context: { credentials: 'one' },
+    }).subscribe(next(data), error, complete);
+    execute(link, {
+      query: sampleQuery,
+      context: { credentials: 'two' },
+    }).subscribe(next(data2), error, complete);
+  });
+
+  it('errors on an incorrect number of results for a batch', done => {
+    const link = new BatchHttpLink({
+      uri: 'batch',
+      batchInterval: 0,
+      batchMax: 3,
     });
+
+    let errors = 0;
+    const next = data => {
+      done.fail('next should not have been called');
+    };
+
+    const complete = () => {
+      done.fail('complete should not have been called');
+    };
+
+    const error = error => {
+      errors++;
+
+      if (errors === 3) {
+        done();
+      }
+    };
+
+    execute(link, { query: sampleQuery }).subscribe(next, error, complete);
+    execute(link, { query: sampleQuery }).subscribe(next, error, complete);
+    execute(link, { query: sampleQuery }).subscribe(next, error, complete);
   });
 });
