+++ packages/apollo-link-batch-http/src/__tests__/sharedHttpTests.ts
@@ -1,0 +1,1 @@
+../../../apollo-link-http/src/__tests__/sharedHttpTests.ts
